# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de LearnUp",
        "description": "Documentación de la API de LearnUp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/institution/": {
            "get": {
                "tags": [
                    "institucion"
                ],
                "summary": "Get institutions",
                "description": "Busqueda por filtros y paginación",
                "operationId": "get Institutions",
                "responses": {
                    "200": {
                        "description": "List of institutions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Univalle"
                                        },
                                        "tipo": {
                                            "type": "string",
                                            "example": "publica"
                                        },
                                        "Capacidad": {
                                            "type": "integer",
                                            "example": 1500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "There are not institutions"
                    }
                }
            }
        },
...
        "/institutions/users/add": {
            "post": {
                "tags": [
                    "institucion"
                ],
                "summary": "Add user to institution",
                "description": "Add a user to the specified institution.",
                "operationId": "addUserToInstitution",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "institucion_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "institucion_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User added successfully"
                    },
                    "400": {
                        "description": "User already in institution"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/institutions/users/remove": {
            "delete": {
                "tags": [
                    "institucion"
                ],
                "summary": "Remove user from institution",
                "description": "Remove a user from the specified institution.",
                "operationId": "removeUserFromInstitution",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "institucion_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "institucion_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User removed successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/institutions/{institucion_id}/users": {
            "get": {
                "tags": [
                    "institucion"
                ],
                "summary": "Get users by institution",
                "description": "Retrieve all users in the specified institution.",
                "operationId": "getUsersByInstitution",
                "parameters": [
                    {
                        "name": "institucion_id",
                        "in": "path",
                        "description": "ID of the institution",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "john@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/institutions/users/check": {
            "post": {
                "tags": [
                    "institucion"
                ],
                "summary": "Check user in institution",
                "description": "Check if a user belongs to the specified institution.",
                "operationId": "isUserInInstitution",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "institucion_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "institucion_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User enrollment status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isEnrolled": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{user_id}/institutions": {
            "get": {
                "tags": [
                    "institucion"
                ],
                "summary": "Get institutions by user",
                "description": "Retrieve all institutions a user belongs to.",
                "operationId": "getInstitutionsByUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of institutions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Univalle"
                                            },
                                            "tipo": {
                                                "type": "string",
                                                "example": "publica"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/institutions/{institucion_id}/students/count": {
            "get": {
                "tags": [
                    "institucion"
                ],
                "summary": "Get number of students in institution",
                "description": "Retrieve the number of students in the specified institution.",
                "operationId": "getNumberOfStudentsByInstitution",
                "parameters": [
                    {
                        "name": "institucion_id",
                        "in": "path",
                        "description": "ID of the institution",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Number of students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 150
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}